Switched from memorysharp to MemoryIO.Pcsx2 Adds support for PCSX2 1.7 + / 1.6 x64
Some code cleanup
Migrated to .net 8.0

+++ Socom2StreamData/Helpers/GeneralFunctionsHelper.cs
@@ -6,7 +6,7 @@ namespace Socom2StreamData
 {
     public static class GeneralFunctionsHelper
     {
-        public static void setLabel(Label label, string playerName, string livingStatus,IntPtr labelName)
+        public static void setLabel(Label label, string playerName, string livingStatus,int labelName)
         {
             label.Text = playerName;
             label.Tag = labelName;
